# backend/proto/absence/v1/absence.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "absence/v1/absence.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/absences": {
      "get": {
        "operationId": "AbsenceService_GetAbsences",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetAbsencesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "tags": [
          "AbsenceService"
        ]
      },
      "post": {
        "operationId": "AbsenceService_CreateAbsence",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateAbsenceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAbsenceRequest"
            }
          }
        ],
        "tags": [
          "AbsenceService"
        ]
      }
    },
...
    "/api/v1/departments": {
      "get": {
        "operationId": "OrganizationService_GetDepartments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetDepartmentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "tags": [
          "OrganizationService"
        ]
      },
      "post": {
        "operationId": "OrganizationService_CreateDepartment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateDepartmentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateDepartmentRequest"
            }
          }
        ],
        "tags": [
          "OrganizationService"
        ]
      }
    },
...
    "/api/v1/holidays": {
      "get": {
        "operationId": "HolidayService_GetHolidays",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetHolidaysResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "HolidayService"
        ]
      },
      "post": {
        "operationId": "HolidayService_CreateHoliday",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateHolidayResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateHolidayRequest"
            }
          }
        ],
        "tags": [
          "HolidayService"
        ]
      }
    },
...
    "/api/v1/teams": {
      "get": {
        "operationId": "OrganizationService_GetTeams",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetTeamsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "departmentId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OrganizationService"
        ]
      },
      "post": {
        "operationId": "OrganizationService_CreateTeam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateTeamResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTeamRequest"
            }
          }
        ],
        "tags": [
          "OrganizationService"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "UserService_GetUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetUsersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "tags": [
          "UserService"
        ]
      },
      "post": {
        "operationId": "UserService_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
...
  },
  "definitions": {
    "Absence": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "reason": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "teamName": {
          "type": "string"
        }
      },
      "title": "Core types"
    },
    "Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "CreateAbsenceRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        },
        "reason": {
          "type": "string"
        },
        "teamName": {
          "type": "string"
        }
      },
      "title": "Absence messages"
    },
    "CreateAbsenceResponse": {
      "type": "object",
      "properties": {
        "absence": {
          "$ref": "#/definitions/Absence"
        }
      }
    },
    "CreateDepartmentRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "title": "Department messages"
    },
    "CreateDepartmentResponse": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        }
      }
    },
    "CreateHolidayRequest": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Holiday messages"
    },
    "CreateHolidayResponse": {
      "type": "object",
      "properties": {
        "holiday": {
          "$ref": "#/definitions/Holiday"
        }
      }
    },
    "CreateTeamRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "departmentId": {
          "type": "string"
        }
      },
      "title": "Team messages"
    },
    "CreateTeamResponse": {
      "type": "object",
      "properties": {
        "team": {
          "$ref": "#/definitions/Team"
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      },
      "title": "User messages"
    },
    "CreateUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "Department": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "GetAbsencesResponse": {
      "type": "object",
      "properties": {
        "absences": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Absence"
          }
        }
      }
    },
    "GetDepartmentsResponse": {
      "type": "object",
      "properties": {
        "departments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Department"
          }
        }
      }
    },
    "GetHolidaysResponse": {
      "type": "object",
      "properties": {
        "holidays": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Holiday"
          }
        }
      }
    },
    "GetTeamsResponse": {
      "type": "object",
      "properties": {
        "teams": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Team"
          }
        }
      }
    },
    "GetUsersResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "Holiday": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Any"
          }
        }
      }
    },
    "Team": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "departmentId": {
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "departmentId": {
          "type": "string"
        },
        "teamId": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "jobProfile": {
          "type": "string"
        }
      }
    }
  }
}